backend/pyproject.toml
[project]
name = "emoji-hero"
version = "0.1.0"
description = "Chat-powered custom emoji generator for Slack"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "opentelemetry-exporter-otlp",
    "opentelemetry-sdk",
    "pillow>=12.1.1",
    "pydantic-ai>=1.70.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "tavily-python>=0.7.23",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]
